Refactor code and add Initializer Interface
